+++ tests/test_02_store_inplace.py
