+++ tests/test_fallback.py
