+++ tests/test_instruction_translator_cache.py
