+++ tests/test_output_restoration.py
