+++ tests/test_resnet50_backward.py
@@ -11,19 +11,55 @@ from numpy.testing import assert_array_equal
 import paddle
 from paddle.vision import resnet50
 from symbolic_trace import symbolic_trace
-from symbolic_trace.symbolic.compile_cache import CompileSIRCache
 from symbolic_trace.utils.utils import execute_time
 
 
-def run_dygraph_optimizer(inp, to_static):
+def resnet_call(net: paddle.nn.Layer, x: paddle.Tensor):
+    return net(x)
+
+
+def run_dygraph_optimizer(inp):
     """dygraph train + SGD optimizer"""
     paddle.seed(2021)
     np.random.seed(2021)
     random.seed(2021)
     net = resnet50()
-    if to_static:
-        net.forward = symbolic_trace(net.forward)
-        # net = paddle.jit.to_static(net)
+    optimizer = paddle.optimizer.SGD(
+        learning_rate=0.03, parameters=net.parameters()
+    )
+    for i in range(5):
+        optimizer.clear_grad()
+        loss = execute_time(net)(inp)
+        loss.backward()
+        optimizer.step()
+    return loss
+
+
+def run_symbolic_optimizer(inp):
+    """dygraph train + SGD optimizer"""
+    paddle.seed(2021)
+    np.random.seed(2021)
+    random.seed(2021)
+    net = resnet50()
+    net_wrapper = symbolic_trace(resnet_call)
+    optimizer = paddle.optimizer.SGD(
+        learning_rate=0.03, parameters=net.parameters()
+    )
+    for i in range(5):
+        optimizer.clear_grad()
+        loss = execute_time(net_wrapper)(net, inp)
+        loss.backward()
+        optimizer.step()
+    return loss
+
+
+def run_to_static_optimizer(inp):
+    """dygraph train + SGD optimizer"""
+    paddle.seed(2021)
+    np.random.seed(2021)
+    random.seed(2021)
+    net = resnet50()
+    net = paddle.jit.to_static(net)
     optimizer = paddle.optimizer.SGD(
         learning_rate=0.03, parameters=net.parameters()
     )
@@ -42,14 +78,15 @@ class TestBackward(unittest.TestCase):
         np.random.seed(2021)
         random.seed(2021)
         inp = paddle.rand((3, 3, 255, 255))
-        out1 = run_dygraph_optimizer(inp, True)[0].numpy()
-        out2 = run_dygraph_optimizer(inp, False)[0].numpy()
+        print("Start Run SymbolicTrace:")
+        out2 = run_symbolic_optimizer(inp)[0].numpy()
+        print("Start Run Dygraph:")
+        out1 = run_dygraph_optimizer(inp)[0].numpy()
+        print("Start Run To Static:")
+        out1 = run_to_static_optimizer(inp)[0].numpy()
         assert_array_equal(
             out1, out2, "Not Equal in dygraph and static graph", True
         )
-        assert (
-            CompileSIRCache().hit_num == 4
-        ), "CompileSIRCache hit_num should be 4"
 
 
 if __name__ == "__main__":
