+++ tests/test_trace_list_arg.py
