+++ tests/tests_legacy/error_test_resnet_with_trace_cache.py
