+++ tests/tests_legacy/error_test_sir_call.py
