+++ tests/tests_legacy/error_test_symbolic_load_before_call.py
