+++ tests/tests_legacy/test_error_cases.py
