+++ tests/tests_legacy/test_symbolic_fallback.py
