+++ tests/tests_legacy/test_symbolic_for.py
