+++ tests/tests_legacy/test_symbolic_gc.py
