+++ tests/tests_legacy/test_symbolic_infer_meta.py
