+++ tests/tests_legacy/test_symbolic_nested.py
