+++ tests/tests_legacy/test_symbolic_trace.py
